requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
duckduckgo-search>=6.0.0
pymupdf>=1.24.0
python-dotenv>=1.0.0